[MRELEASE-1056] Rename package to org.apache.maven.shared.release.transform.jdom2
