+++ maven-release-manager/src/main/java/org/apache/maven/shared/release/transform/jdom2/JDomModelBase.java
@@ -1,4 +1,4 @@
-package org.apache.maven.shared.release.transform.jdom;
+package org.apache.maven.shared.release.transform.jdom2;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
