+++ maven-release-manager/src/test/java/org/apache/maven/shared/release/transform/jdom2/JDomScmTest.java
@@ -1,4 +1,4 @@
-package org.apache.maven.shared.release.transform.jdom;
+package org.apache.maven.shared.release.transform.jdom2;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
